requirements.txt
influxdb-client==1.21.0
pandas==1.3.3
numpy==1.21.2